+++ activity.go
@@ -80,7 +80,7 @@ type Activity struct {
 	Type      ActivityType           `json:"type"`
 
 	// The identity that performed this activity.
-	Actor IdentityRef `json:"actor"`
+	Actor *IdentityRef `json:"actor,omitempty"`
 
 	// Optional references to documents that were affected by the activity.
 	Identities               []IdentityRef                `json:"identities,omitempty"`
@@ -97,6 +97,20 @@ type Activity struct {
 	// Session and request IDs from the WAF framework.
 	SessionID identifier.Identifier `json:"sessionId"`
 	RequestID identifier.Identifier `json:"requestId"`
+}
+
+func (a *Activity) IsForOrganization(organization *Organization) bool {
+	// Check if activity includes only one organization.
+	if len(a.Organizations) != 1 {
+		return false
+	}
+
+	// Check if activity includes this organization.
+	if a.Organizations[0].ID != *organization.ID {
+		return false
+	}
+
+	return true
 }
 
 func (a *Activity) Validate(_ context.Context, existing *Activity) errors.E {
@@ -154,6 +168,15 @@ func (s *Service) getActivity(_ context.Context, id identifier.Identifier) (*Act
 	return &activity, nil
 }
 
+func (s *Service) getActivityFromID(ctx context.Context, value string) (*Activity, errors.E) {
+	id, errE := identifier.MaybeString(value)
+	if errE != nil {
+		return nil, errors.WrapWith(errE, ErrActivityNotFound)
+	}
+
+	return s.getActivity(ctx, id)
+}
+
 func (s *Service) createActivity(ctx context.Context, activity *Activity) errors.E {
 	errE := activity.Validate(ctx, nil)
 	if errE != nil {
@@ -195,7 +218,7 @@ func (s *Service) logActivity(
 		Timestamp: Time{}, //nolint:exhaustruct
 
 		Type:                     activityType,
-		Actor:                    IdentityRef{ID: actorID},
+		Actor:                    &IdentityRef{ID: actorID},
 		Providers:                providers,
 		Changes:                  changes,
 		SessionID:                sessionID,
@@ -278,13 +301,7 @@ func (s *Service) ActivityGetGet(w http.ResponseWriter, req *http.Request, param
 		return
 	}
 
-	activityID, errE := identifier.MaybeString(params["id"])
-	if errE != nil {
-		s.NotFoundWithError(w, req, errE)
-		return
-	}
-
-	activity, errE := s.getActivity(ctx, activityID)
+	activity, errE := s.getActivityFromID(ctx, params["id"])
 	if errors.Is(errE, ErrActivityNotFound) {
 		s.NotFoundWithError(w, req, errE)
 		return
@@ -301,6 +318,155 @@ func (s *Service) ActivityGetGet(w http.ResponseWriter, req *http.Request, param
 		waf.Error(w, req, http.StatusUnauthorized)
 		return
 	}
+
+	s.WriteJSON(w, req, activity, nil)
+}
+
+func (s *Service) OrganizationActivity(w http.ResponseWriter, req *http.Request, _ waf.Params) {
+	if s.ProxyStaticTo != "" {
+		s.Proxy(w, req)
+	} else {
+		s.ServeStaticFile(w, req, "/index.html")
+	}
+}
+
+func (s *Service) OrganizationActivityGet(w http.ResponseWriter, req *http.Request, params waf.Params) {
+	ctx := s.RequireAuthenticated(w, req)
+	if ctx == nil {
+		return
+	}
+
+	organization, errE := s.getOrganizationFromID(ctx, params["id"])
+	if errors.Is(errE, ErrOrganizationNotFound) {
+		s.NotFoundWithError(w, req, errE)
+		return
+	} else if errE != nil {
+		s.InternalServerErrorWithError(w, req, errE)
+		return
+	}
+
+	if !organization.HasAdminAccess(IdentityRef{ID: mustGetIdentityID(ctx)}) {
+		waf.Error(w, req, http.StatusUnauthorized)
+		return
+	}
+
+	result := []ActivityRef{}
+
+	s.activitiesMu.RLock()
+	defer s.activitiesMu.RUnlock()
+
+	// Collect activities that include this organization exactly (not spanning multiple organizations).
+	activities := make([]*Activity, 0)
+	for id, data := range s.activities {
+		var activity Activity
+		errE := x.UnmarshalWithoutUnknownFields(data, &activity)
+		if errE != nil {
+			errors.Details(errE)["id"] = id
+			s.InternalServerErrorWithError(w, req, errE)
+			return
+		}
+
+		if !activity.IsForOrganization(organization) {
+			continue
+		}
+
+		activities = append(activities, &activity)
+	}
+
+	// Sort activities by timestamp (newest first).
+	slices.SortFunc(activities, func(a, b *Activity) int {
+		return time.Time(b.Timestamp).Compare(time.Time(a.Timestamp))
+	})
+
+	// Convert to refs.
+	for _, activity := range activities {
+		result = append(result, ActivityRef{ID: *activity.ID})
+	}
+
+	s.WriteJSON(w, req, result, nil)
+}
+
+func (s *Service) OrganizationActivityGetGet(w http.ResponseWriter, req *http.Request, params waf.Params) {
+	ctx := s.RequireAuthenticated(w, req)
+	if ctx == nil {
+		return
+	}
+
+	organization, errE := s.getOrganizationFromID(ctx, params["id"])
+	if errors.Is(errE, ErrOrganizationNotFound) {
+		s.NotFoundWithError(w, req, errE)
+		return
+	} else if errE != nil {
+		s.InternalServerErrorWithError(w, req, errE)
+		return
+	}
+
+	if !organization.HasAdminAccess(IdentityRef{ID: mustGetIdentityID(ctx)}) {
+		waf.Error(w, req, http.StatusUnauthorized)
+		return
+	}
+
+	activity, errE := s.getActivityFromID(ctx, params["activityId"])
+	if errors.Is(errE, ErrActivityNotFound) {
+		s.NotFoundWithError(w, req, errE)
+		return
+	} else if errE != nil {
+		s.InternalServerErrorWithError(w, req, errE)
+		return
+	}
+
+	// Verify this activity is for this organization.
+	if !activity.IsForOrganization(organization) {
+		s.NotFound(w, req)
+		return
+	}
+
+	// Map actor identity ID to organization-scoped ID.
+	actorIdentity, errE := s.getIdentityWithoutAccessCheck(ctx, activity.Actor.ID)
+	if errors.Is(errE, ErrActivityNotFound) {
+		// It should not really happen that we cannot get the identity because
+		// currently we do not support deleting identities.
+		// TODO: Once we do support deleting identities AND have identity refs with basic information, we should remove just the ID and keep basic information.
+		activity.Actor = nil
+	} else if errE != nil {
+		s.InternalServerErrorWithError(w, req, errE)
+		return
+	}
+	idOrg := actorIdentity.GetOrganization(organization.ID)
+	if idOrg == nil || !idOrg.Active {
+		// TODO: Once we have identity refs with basic information, we should remove just the ID and keep basic information.
+		activity.Actor = nil
+	} else {
+		activity.Actor.ID = *idOrg.ID
+	}
+
+	// Map identity IDs in the Identities slice to organization-scoped IDs.
+	scopedIdentities := []IdentityRef{}
+	for _, identityRef := range activity.Identities {
+		identity, errE := s.getIdentityWithoutAccessCheck(ctx, identityRef.ID)
+		if errors.Is(errE, ErrActivityNotFound) {
+			// It should not really happen that we cannot get the identity because
+			// currently we do not support deleting identities.
+			// TODO: Once we do support deleting identities AND have identity refs with basic information, we should remove just the ID and keep basic information.
+			continue
+		} else if errE != nil {
+			s.InternalServerErrorWithError(w, req, errE)
+			return
+		}
+		idOrg := identity.GetOrganization(organization.ID)
+		if idOrg == nil || !idOrg.Active {
+			// TODO: Once we have identity refs with basic information, we should remove just the ID and keep basic information.
+			continue
+		}
+
+		identityRef.ID = *idOrg.ID
+		scopedIdentities = append(scopedIdentities, identityRef)
+	}
+
+	if len(scopedIdentities) == 0 {
+		scopedIdentities = nil
+	}
+	activity.Identities = scopedIdentities
 
 	s.WriteJSON(w, req, activity, nil)
 }