+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internationalization support with English and Slovenian locales.
   [#8](https://gitlab.com/charon/charon/-/issues/8)
 - Activity log for users to see actions related to their identities.
+- Activity log for organization admins to see actions related to their organizations.
 
 ## [0.4.0] - 2025-06-16